requirements.txt
google-cloud-texttospeech
google-cloud-translate
langchain
langchain-openai
openai
pydub
python-dotenv
requests
telegram
telebot